mv userCard in it's component
