+++ frontend/js/components/userTemplate/userTemplate.js
@@ -10,7 +10,3 @@ export async function userTemplateComponent() {
         console.error('Error fetching userTemplateComponent:', error);
     }
 }
-
-export async function userTemplateListener() {
-
-}
